conn: move ConnAccount functions to libconn
These functions no longer have backend dependencies, so move them to the
connection library.

- mutt_account_getlogin
- mutt_account_getoauthbearer
- mutt_account_getpass
- mutt_account_getuser
- mutt_account_unsetpass
