+++ imap/auth_oauth.c
@@ -33,6 +33,7 @@
 #include "mutt/lib.h"
 #include "conn/lib.h"
 #include "auth.h"
+#include "lib.h"
 #include "mutt_account.h"
 #include "mutt_logging.h"
 #include "mutt_socket.h"
