+++ mutt_account.h
@@ -31,18 +31,6 @@
 struct ConnAccount;
 struct Url;
 
-/* These Config Variables are only used in mutt_account.c */
-extern char *C_ImapLogin;
-extern char *C_ImapOauthRefreshCommand;
-extern char *C_ImapPass;
-extern char *C_NntpPass;
-extern char *C_NntpUser;
-extern char *C_PopOauthRefreshCommand;
-extern char *C_PopPass;
-extern char *C_PopUser;
-extern char *C_SmtpOauthRefreshCommand;
-extern char *C_SmtpPass;
-
 /**
  * enum AccountType - account types
  */
@@ -55,20 +43,7 @@ enum AccountType
   MUTT_ACCT_TYPE_NNTP,     ///< Nntp (Usenet) Account
 };
 
-typedef uint8_t MuttAccountFlags;     ///< Flags, Which ConnAccount fields are initialised, e.g. #MUTT_ACCT_PORT
-#define MUTT_ACCT_NO_FLAGS        0   ///< No flags are set
-#define MUTT_ACCT_PORT      (1 << 0)  ///< Port field has been set
-#define MUTT_ACCT_USER      (1 << 1)  ///< User field has been set
-#define MUTT_ACCT_LOGIN     (1 << 2)  ///< Login field has been set
-#define MUTT_ACCT_PASS      (1 << 3)  ///< Password field has been set
-#define MUTT_ACCT_SSL       (1 << 4)  ///< Account uses SSL/TLS
-
-int   mutt_account_fromurl       (struct ConnAccount *account, const struct Url *url);
-void  mutt_account_tourl         (struct ConnAccount *account, struct Url *url);
-int   mutt_account_getuser       (struct ConnAccount *account);
-int   mutt_account_getlogin      (struct ConnAccount *account);
-int   mutt_account_getpass       (struct ConnAccount *account);
-void  mutt_account_unsetpass     (struct ConnAccount *account);
-char *mutt_account_getoauthbearer(struct ConnAccount *account);
+int   mutt_account_fromurl(struct ConnAccount *account, const struct Url *url);
+void  mutt_account_tourl  (struct ConnAccount *account, struct Url *url);
 
 #endif /* MUTT_MUTT_ACCOUNT_H */
