+++ pop/pop_auth.c
@@ -35,6 +35,7 @@
 #include "mutt/lib.h"
 #include "address/lib.h"
 #include "conn/lib.h"
+#include "lib.h"
 #include "mutt_account.h"
 #include "mutt_logging.h"
 #include "mutt_socket.h"
