+++ smtp.h
@@ -29,6 +29,9 @@
 
 /* These Config Variables are only used in smtp.c */
 extern struct Slist *C_SmtpAuthenticators;
+extern char *C_SmtpOauthRefreshCommand;
+extern char *C_SmtpPass;
+extern char *C_SmtpUser;
 
 #ifdef USE_SMTP
 struct AddressList;
